Create genmd pkg
