+++ generatedocs/cmd/generatemarkdown/generate.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/qlik-oss/gopherciser/generatedocs/pkg/genmd"
+
+func main() {
+	genmd.GenerateMarkdown()
+}
