+++ main.go
@@ -5,10 +5,10 @@ import (
 )
 
 // Compile documentation data to be used by GUI and for markdown generation
-//go:generate go run ./generatedocs/cmd/compiledocs/main.go
+//go:generate go run ./generatedocs/cmd/compiledocs
 
 // Generate markdown files
-//go:generate go run ./generatedocs/generate/generate.go --output ./docs/settingup.md
+//go:generate go run ./generatedocs/cmd/generatemarkdown --output ./docs/settingup.md
 
 func main() {
 	cmd.Execute()
